Bump hdrsample to 5.0
This requires using `Histogram`s with `u64` counters instead of `usize`
counters since hdrsample removed the ability to use a usize counter type
in 4.0, but it's more correct to use a u64, anyway. 32 bit counters are
an overflow risk.

+++ src/prometheus.rs
@@ -44,7 +44,7 @@ fn write_buckets<N, W>(
     out: &mut W,
     name: &N,
     labels: &FmtLabels,
-    h: &Histogram<usize>,
+    h: &Histogram<u64>,
 ) -> fmt::Result
 where
     N: fmt::Display,
@@ -80,7 +80,7 @@ fn write_bucket<N, M, W>(
     name: &N,
     labels: &FmtLabels,
     le: &M,
-    count: usize,
+    count: u64,
 ) -> fmt::Result
 where
     N: fmt::Display,
